Avoid lazy initialization when obtaining persistent class
Current utility code in Hibernate class forces initialization of the
proxies just to obtain the target persistent class that is used, because
it uses `getImplementation()` method of LazyInitializer.

There is a method in LazyInitializer that can do the same without
initialization (`getPersistentClass()`), so I think it should be used instead.

I created HHH-13857 to track this.